+++ src/app/actions/user.js
@@ -1,0 +1,9 @@
+'use strict';
+
+import Reflux from 'reflux';
+
+const UserActions = Reflux.createActions({
+	notifyQuota: {}
+});
+
+module.exports = UserActions;