+++ src/app/libs/notify-core.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import {upperFirst, deepAssign, isEqual} from 'app/libs/helpers';
+import {upperFirst, deepAssign, isEqual, isObject} from 'app/libs/helpers';
 
 const debug = require('debug')('AiC:Libs:NotifyCore');
 
@@ -37,18 +37,22 @@ class NotifyCore {
 	clearGroup(group, groupInfo) {
 		debug(`clear ${group}`, groupInfo);
 		const groupId = this.getGroupId(group, groupInfo);
-		const groupIdName = this.groups[group].id;
+		// const groupIdName = this.groups[group].id;
 		this.groups[group].activeCount[groupId]--;
 
 		// Nobody is watching this group, all calls should be removed
 		if (this.groups[group].activeCount[groupId] === 0) {
 			this.groups[group].children.forEach(action => {
-				this.actions[action].running
-					.filter(r => r.actionInfo[groupIdName] === groupId)
-					.forEach(r => {
-						r.shouldStop = true;
-						clearTimeout(r.timeout);
-					});
+				if (this.isActionRunning(action, groupInfo)) {
+					this.actions[action].running[groupId].shouldStop = true;
+					clearTimeout(this.actions[action].running[groupId].timeout);
+				}
+				// this.actions[action].running[groupId]
+					// .filter(r => r.actionInfo[groupIdName] === groupId)
+					// .forEach(r => {
+					// 	r.shouldStop = true;
+					// 	clearTimeout(r.timeout);
+					// });
 			});
 		}
 		debug(`end clear ${group}`, this.groups, this.actions);
@@ -68,12 +72,12 @@ class NotifyCore {
 
 		this.actions[action] = deepAssign({}, actionOptions);
 		this.actions[action].action = action;
-		this.actions[action].running = [];
+		this.actions[action].running = {};
 		const actionLabel = upperFirst(action);
 
 		// Register start and stop functions
 		this[`start${actionLabel}`] = this.startAction.bind(this, action);
-		this[`stop${actionLabel}`] = this.stopAction.bind(this, action);
+		// this[`stop${actionLabel}`] = this.stopAction.bind(this, action);
 
 		// Register in the group
 		if ('group' in actionOptions) {
@@ -81,21 +85,31 @@ class NotifyCore {
 		}
 	}
 
+	isActionRunning(action, actionInfo) {
+		const groupId = this.getGroupIdByAction(action, actionInfo);
+		return groupId in this.actions[action].running && !this.actions[action].running[groupId].shouldStop;
+	}
+
 	startAction(action, actionInfo) {
 		debug(`start${action}`, actionInfo);
 		// TODO should reject calls with same action and actionInfo
+		if (this.isActionRunning(action, actionInfo)) {
+			debug(`Start ${action} ignored. Already running.`, actionInfo);
+			return;
+		}
 
 		// Should only start action if the group is being watched
 		if (this.isActionWatched(action, actionInfo)) {
 			const initialDelaySeconds = 0;
 			const timeoutSeconds = 5;
-			const runningIndex = this.actions[action].running.length;
-			this.actions[action].running.push({shouldStop: false, lastResponse: null, actionInfo});
+			// const runningIndex = this.actions[action].running.length;
+			const runningIndex = this.getGroupIdByAction(action, actionInfo);
+			this.actions[action].running[runningIndex] = {shouldStop: false, lastResponse: null, actionInfo};
 			const fn = () => {
 				debug(`fn ${action}`, actionInfo);
 				this.actions[action].request(actionInfo).then(response => {
 					debug(`response ${action}`, response);
-					const finalResponse = 'response' in response ? response.response : response;
+					const finalResponse = isObject(response) && 'response' in response ? response.response : response;
 					if (!isEqual(finalResponse, this.actions[action].running[runningIndex].lastResponse)) {
 						debug(`not equal, notifying ${action}`);
 						this.actions[action].running[runningIndex].lastResponse = finalResponse;
@@ -110,14 +124,14 @@ class NotifyCore {
 			};
 			this.actions[action].running[runningIndex].timeout = setTimeout(fn, initialDelaySeconds * 1000);
 		} else {
-			debug(`start ${action} ignored. Nobody watching.`, actionInfo);
+			debug(`Start ${action} ignored. Nobody watching.`, actionInfo);
 		}
 	}
 
-	stopAction(action, actionInfo) {
-		debug(`stop${action}`, actionInfo);
-		// this.actions[action].running[runningIndex].shouldStop = true;
-	}
+	// stopAction(action, actionInfo) {
+	// 	debug(`stop${action}`, actionInfo);
+	// 	// this.actions[action].running[runningIndex].shouldStop = true;
+	// }
 
 	registerActions(info) {
 		const keys = Object.keys(info);
@@ -133,10 +147,19 @@ class NotifyCore {
 		});
 	}
 
+	getGroupIdByAction(action, actionInfo) {
+		const group = this.getGroupByAction(action);
+		return this.getGroupId(group, actionInfo);
+	}
+
+	getGroupByAction(action) {
+		return this.actions[action].group;
+	}
+
 	isActionWatched(action, actionInfo) {
 		debug(`isActionWatched ${action}`, actionInfo);
-		const group = this.actions[action].group;
-		const groupId = this.getGroupId(group, actionInfo);
+		const group = this.getGroupByAction(action);
+		const groupId = this.getGroupIdByAction(action, actionInfo);
 		debug(`isActionWatched vars`, group, groupId, this.groups[group].activeCount[groupId]);
 		debug(`isActionWatched class`, this.groups, this.actions);
 		if (!this.actionExists(action) ||